Split the views in composable views
